requirements.txt
streamlit
openai
python-dotenv
fastapi
uvicorn
flask
python-multipart
python-docx
PyPDF2
langchain
